requirements.txt
asgiref==3.3.0
astor==0.8.1
attrs==20.3.0
bandit==1.6.2
darglint==1.5.5
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
docutils==0.16
eradicate==1.0
factory-boy==3.1.0
Faker==4.14.2
flake8==3.8.4
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-commas==2.0.0
flake8-comprehensions==3.3.0
flake8-debugger==3.2.1
flake8-docstrings==1.5.0
flake8-eradicate==0.3.0
flake8-isort==3.0.1
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
gitdb==4.0.5
GitPython==3.1.11
inflection==0.5.1
iniconfig==1.1.1
isort==4.3.21
mccabe==0.6.1
packaging==20.4
pbr==5.5.1
pep8-naming==0.9.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytest-factoryboy==2.0.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
restructuredtext-lint==1.3.1
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
sqlparse==0.4.1
stevedore==3.2.2
testfixtures==6.15.0
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
wemake-python-styleguide==0.14.1
